commot031121

+++ stk_header.h
@@ -14,9 +14,10 @@
 #define CHECK_STK_  check_stk_err (stk, __LINE__, __func__)
 
 typedef int data_type;
-typedef uint64_t can_type;
+typedef int can_type;
 
 const int BASE_STK_SIZE = 2;
+
 const can_type const_r_stk_canary = 0xDEDADEAD;
 const can_type const_l_stk_canary = 0xDEADBABA;
 
@@ -26,6 +27,7 @@ enum resize_mode {
 };
 
 enum errors {
+
     NO_ERROR = 0,
     STACK_UNDERFLOW       = -1,
     STACK_OVERFLOW        = -2,
@@ -38,26 +40,34 @@ enum errors {
     STK_RIGHT_CANARY_ERR  = -9, 
     HASH_ERR              = -10,
     RESIZE_ERR            = -11
+
 };
 
 typedef struct Stack_tag {
     #ifndef NO_PROTECT
+    
         can_type left_stk_can;
+    
     #endif
 
     #ifdef TWO_LVL_PROTECT
+
         int hash;
+    
     #endif
 
     char* data_canary;
     data_type* data;
+
     int size;
     int capacity;
 
-
     #ifndef NO_PROTECT
+
         can_type right_stk_can;
+    
     #endif
+
 } Stack;
 
 void       StackPush           (Stack *stk, const data_type value);
